Run the Full Inspections Table Pipeline

This is a simplified wrapper for the inspections table pipeline that makes it easier to run. This takes inspections data from load to finished product and cleans up after itself. See the component functions for details on the pipeline.

insp_table_pipeline(
  data_path = coviData::path_create("V:/Compliance/Inspection Data for Publishing/",
    "Grand List of Inspections", ext = "xlsx"),
 
    table_path = coviData::path_create("V:/Compliance/Inspection Data for Publishing/Table/",
    paste0("insp_table_", Sys.Date()), ext = "html"),
  archive_dir = "V:/Compliance/Inspection Data for Publishing/Table/Archive/",
  force = FALSE
)

Arguments

data_path

The path to the excel workbook containing inspections data

table_path

The location to save the inspections table

archive_dir

The path to the archive directory for inspections tables

force

Should existing files at table_path or archive_dir be overwritten if a conflict occurs?

Value

The inspections table as a datatables object (invisibly)

See also