Create a Location for a Flextable into a Powerpoint

Create a Location for a Flextable into a Powerpoint

ph_location_table(
  flextable,
  pptx,
  layout,
  master = NULL,
  type = "tbl",
  pos_h = TRUE,
  pos_first = TRUE,
  halign = 0.5,
  valign = 0.5,
  newlabel = "",
  id = 1
)

Arguments

flextable
The flextable to add

pptx
The powerpoint object

layout
The layout of the slide being added to

master
The slide master to use, if any

type
The placeholder type to use

pos_h
If multiple of the same placeholder type exist, how should the placeholder be chosen? TRUE means the horizontal axis will be used; FALSE means vertical. See pos_first for more information.

pos_first
If multiple of the same placeholder type exist, should the first (or last) be chosen (along the axis given by pos_h)?

halign
Horizontal table alignment with placeholder; 0 aligns left borders, 1 aligns right borders. Defaults to centered (0.5).

valign
Vertical table alignment with placeholder; 0 aligns bottom borders, 1 aligns top borders. Defaults to centered (0.5).

newlabel
a label for the placeholder. See section details.

id
index of the placeholder template. If two body placeholder, there can be two different index: 1 and 2 for the first and second body placeholders defined in the layout.