Plot Vaccinations by Age

plot_vac_age(
  .data = coviData::vac_prep(coviData::read_vac()),
  pct = TRUE,
  by_pop = TRUE,
  incl_under_15 = FALSE,
  incl_under_12 = incl_under_15
)

Arguments

.data

Vaccination data, as created by vac_prep()

pct

Deprecated.

by_pop

Should the resulting graphic be calculated using age group population statistics (TRUE, the default) or not (FALSE)?

incl_under_15

Deprecated.

incl_under_12

Should the age group "0-11" be included? The default is FALSE.

Value

A ggplot object

Details

This function is deprecated. Please use the vac_plot_age() function, which is now the backend for this function.