Plot Vaccinations by Age Group

vac_plot_age() plots vaccinations by age group in a bar chart.

vac_plot_age(
  .data = vac_prep(read_vac(date)),
  date = NULL,
  by_pop = TRUE,
  incl_under_12 = FALSE
)

Arguments

.data

Vaccination data, as created by vac_prep()

date

Date of vaccination data

by_pop

Should the resulting graphic be calculated using age group population statistics (TRUE, the default) or not (FALSE)?

incl_under_12

Should the age group "0-11" be included? The default is FALSE.

Value

A ggplot object

Details

When by_pop = TRUE, the resulting figure displays vaccinations with reference to the population of each age group; when by_pop = FALSE, the figure does not display population-related values.